Refactor: Separate NIP-65 relay list functionality from contact management
This commit separates the NIP-65 relay list management from the NIP-02 contact list management as requested.

A new `relay` command has been created to handle all relay list operations (set, get, edit). The existing `contact` command is now solely responsible for contact list operations (add, list).

This improves the code's organization and clarity by separating distinct functionalities into their own modules and commands.
